+++ examples/00-simple-single-page/fluxexapp.js
@@ -1,6 +1,6 @@
 'use strict';
 
-require('fluxex/extra/polyfill');
+require('babel-polyfill');
 
 module.exports = require('fluxex').createApp({
     sampleStore: require('./stores/sample'),