+++ extra/polyfill.js
@@ -1,4 +1,0 @@
-// This is a dummy file to enable babel-core polyfill for browser
-// Do this in your first line of fluxex.js to enable polyfill for browsers:
-// require('fluxex/extra/polyfill');
-// our gulp task will alias it to 'babelify/polyfill'