+++ starter.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
-npm install --save react babel express routr request body-parser babel-runtime
-npm install --save-dev browser-request aliasify browserify watchify babelify envify eslint eslint-plugin-react babel-eslint nodemon browser-sync gulp gulp-babel gulp-cached gulp-jsx-coverage gulp-eslint gulp-util gulp-uglify vinyl-source-stream vinyl-buffer
+npm install --save react babel express routr request body-parser
+npm install --save-dev browser-request aliasify browserify watchify babelify envify eslint eslint-plugin-react babel-eslint nodemon browser-sync gulp gulp-babel gulp-cached gulp-jsx-coverage gulp-eslint gulp-util gulp-uglify vinyl-source-stream vinyl-buffer babel-polyfill babel-preset-es2015 babel-preset-react babel-plugin-transform-runtime
 mkdir actions
 mkdir components
 mkdir stores
@@ -12,6 +12,7 @@ cp $BASE/server.js index.js
 cp $BASE/fluxexapp.js fluxexapp.js
 echo "require('fluxex/extra/gulpfile');" > gulpfile.js
 cp $BASE/.eslintrc .
+cp $BASE/.babelrc .
 
 echo '==================== THE STARTER PROJECT IS READY ===================='
 echo '= + index.js                    Your Server, add api URL here        ='